requirements.txt
python-dotenv
sqlalchemy
fastapi[all]
psycopg2-binary
instructor
apscheduler